Move stuff around to make room for language server.
