+++ NuGet.config
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<configuration>
+  <packageSources>
+  	<add key="nuget.org" value="https://api.nuget.org/v3/index.json" protocolVersion="3" />
+    <add key="OmniSharp.LSP" value="https://ci.appveyor.com/nuget/csharp-language-server-protocol" />
+  </packageSources>
+</configuration>
