+++ package.json
@@ -2,7 +2,7 @@
     "name": "msbuild-project-tools",
     "displayName": "MSBuild project tools",
     "description": "Tools for working with MSBuild project files (such as auto-complete for package Ids / versions).",
-    "version": "0.0.2",
+    "version": "0.0.3",
     "publisher": "tintoy",
     "license": "MIT",
     "repository": {
@@ -27,7 +27,7 @@
         "workspaceContains:NuGet.config",
         "onLanguage:xml"
     ],
-    "main": "./out/src/extension",
+    "main": "./out/src/extension/extension",
     "contributes": {
         "snippets": [
             {
