+++ src/extension/providers/package-reference-completion.ts
