+++ test/extension/extension.test.ts
@@ -9,7 +9,7 @@ import * as assert from 'assert';
 // You can import and use all API from the 'vscode' module
 // as well as import your extension to test it
 import * as vscode from 'vscode';
-import * as packageReferenceExtension from '../src/extension';
+import * as packageReferenceExtension from '../../src/extension/extension';
 
 // Defines a Mocha test suite to group tests of similar kind together
 suite('Extension Tests', () => {
